Update for IPython 3
