+++ README.md
@@ -1,4 +1,17 @@
-ipython-jsobject
-================
+# ipython-jsobject
 
 Exposes the JS objects directly in Python.  Transparent layer that allows JS to communicate to Python and vise-versa
+
+## Installation
+
+Clone the repository and cd into it, then run:
+
+```
+pip install .
+```
+
+For a development install, run:
+
+```
+pip install -e .
+```
